[118]: Refac > Extract function
