+++ functions/index.js
@@ -3,7 +3,6 @@ const { sendEmail } = require('./emails')
 const { PubSub } = require('@google-cloud/pubsub')
 const admin = require('firebase-admin')
 const { addDiscordRole } = require('./discord_integration')
-const { ethers } = require("ethers")
 const { userCompletedCourse } = require('./checkUserLessons')
 
 admin.initializeApp()
@@ -99,26 +98,11 @@ exports.mintNFT = functions.firestore
     const cohort = (await db.collection('cohorts').doc(createdLesson.cohort_id).get()).data()
     const course = (await db.collection('courses').doc(cohort.data().course_id).get()).data()
     
-    if(!userCompletedCourse(user.id, cohort.data().course_id, db))
+    if(!userCompletedCourse(user.id, cohort.course_id, db))
       return console.log('Usuário não completou todas as lições')
       
-    const contractABI = require('../nfts/artifacts/contracts/BootcampNFTContract.sol/W3DBootcamp.json')
-    const contractAddress = '0xeA4aDb7A75a2452585204E8B411c64C2f3693845'
-    const provider = new ethers.providers.JsonRpcProvider(process.env.ALCHEMY_API_KEY)
-    const signer = new ethers.Wallet( process.env.PRIVATE_KEY, provider)
-    const nftContract = new ethers.Contract(contractAddress, contractABI.abi, signer)
-    await nftContract.mintCertificate(createdLesson.cohort_id, course.nft_title, user.wallet)
-    nftContract.once('Transfer', async (a, b, id) => {
-      console.log('new id ' + id + ' for ' + b)
-      const params = {
-        cohort,
-        course_title: course.nft_title,
-        wallet_address: user.data().wallet,
-        nft_contract: contractAddress,
-        nft_id: id,
-      }
-      await sendEmail('nft_delivery.js', '👷👷‍♀️ WEB3DEV - NFT Recebido: Smart Contract Solidity', user.data().email, params)
-    })
+    mint((createdLesson.cohort_id, course.nft_title, user.wallet))
+
   })
 
 exports.sendEmailJob = functions.pubsub.topic("course_day_email").onPublish((message) => {
